requirements.txt
scapy==2.5.0
requests==2.31.0
beautifulsoup4==4.12.2
pyOpenSSL==23.2.0
PyQt5==5.15.9
fpdf2==2.7.6
python-nmap==0.7.1
cryptography==41.0.3
dnspython==2.4.2 